Added PowerPills; Added Ghost Eatable Animations; Added Ghost Script
Added PowerPills and their base logic: They give you points and make a call to the ghost controller which sets all ghosts to eatable (currently forerver)

Added a Ghost Script which know has all ghost logic  in it that does not belong to movement (e.g. the OnTriggerEnter function to determine who to react if pacman entered the ghost)

Added the needed Animations and a ghost vulnerable (eatable; might get changed) override animation controller: The override controller uses the same logic as the standard ghost animation controller, but with the new ghost animations for visualization of being eatable.