Changed Dot Logic; Removed PowerPill Script; Added Portal Spawn Logic
There is now only one script for both Dots and PowerPills. They now call a Method on the Pacman Script which handles everything.

As a result the PowerPill script was removed.

After the last child has been eaten from the Dots Object it spawns a Prefab on a given position. This will later be used to spawn a portal which Pacman needs to enter to get to the next stage.

The initial Ghost isMovable State was changed from true to false, due to problems with a 4th ghost. It is also a more logical state for the start because the state is set to false in the beginning.